modified process of DeleteFlash

+++ rn/rn_app/src/containers/flashs/ShowFlash.tsx
@@ -1,4 +1,5 @@
 import React, {useRef, useEffect} from 'react';
+import {Alert} from 'react-native';
 import {useDispatch} from 'react-redux';
 import {useNavigation} from '@react-navigation/native';
 import {StackNavigationProp} from '@react-navigation/stack';
@@ -11,6 +12,7 @@ import {RootState, AppDispatch} from '../../redux/index';
 import {selectAllFlashes} from '../../redux/flashes';
 import {deleteFlashThunk} from '../../actions/flashes';
 import {flashMessage} from '../../helpers/flashMessage';
+import {alertSomeError} from '../../helpers/error';
 
 type RootNavigationProp = StackNavigationProp<RootStackParamList, 'ShowFlash'>;
 
@@ -38,8 +40,32 @@ export const Container = ({route}: Props) => {
   const dispatch: AppDispatch = useDispatch();
 
   const deleteFlash = async ({flashId}: {flashId: number}) => {
-    await dispatch(deleteFlashThunk({flashId}));
-    flashMessage('削除しました', 'success');
+    Alert.alert('本当に削除してもよろしいですか?', '', [
+      {
+        text: 'はい',
+        onPress: async () => {
+          const result = await dispatch(deleteFlashThunk({flashId}));
+          if (deleteFlashThunk.fulfilled.match(result)) {
+            flashMessage('削除しました', 'success');
+          } else {
+            if (result.payload && result.payload.errorType === 'invalidError') {
+              flashMessage(result.payload.message, 'danger');
+            } else if (
+              result.payload &&
+              result.payload.errorType === 'someError'
+            ) {
+              alertSomeError();
+            }
+          }
+        },
+      },
+      {
+        text: 'いいえ',
+        onPress: () => {
+          return;
+        },
+      },
+    ]);
   };
 
   return (
